% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Coefficients of a linreg object}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{linreg}.}

\item{...}{Further arguments (not used).}
}
\value{
A named numeric vector of regression coefficients.
}
\description{
Coefficients of a linreg object
}
\examples{
mod <- linreg(Petal.Length ~ Species, data = iris)
coef(mod)
}
