% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linreg.R
\name{linreg_fit}
\alias{linreg_fit}
\title{Fit a linear regression using OLS (wrapper around the RC class)}
\usage{
linreg_fit(formula, data)
}
\arguments{
\item{formula}{A formula like \code{y ~ x1 + x2}.}

\item{data}{A \code{data.frame} containing the variables.}
}
\value{
An object of class \code{"linreg"} (RC) with fields and methods.
}
\description{
This convenience function constructs and returns an object of class \strong{"linreg"}
(the Reference Class you defined) by calling its generator with \verb{$new()}.
}
\examples{
data(iris)
m <- linreg_fit(Petal.Length ~ Sepal.Width + Sepal.Length, iris)
m$summary()
}
