% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Predicted values of a linreg object}
\usage{
\method{pred}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{linreg}.}

\item{...}{Further arguments (not used).}
}
\value{
A numeric vector of fitted values.
}
\description{
Returns the fitted values \eqn{\hat{y} = X\hat{\beta}}.
}
\examples{
mod <- linreg(Petal.Length ~ Species, data = iris)
head(pred(mod))
}
