% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pred}
\alias{pred}
\title{Predicted values}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{A model object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A numeric vector of predicted values.
}
\description{
Generic function returning the predicted (fitted) values of a model.
}
