% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.linreg}
\alias{print.linreg}
\title{Print a linreg object}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{linreg}.}

\item{...}{Further arguments (not used).}
}
\value{
\code{x}, invisibly.
}
\description{
Prints the call and the estimated coefficients with their names,
similar to \code{print.lm}.
}
\examples{
mod <- linreg(Petal.Length ~ Species, data = iris)
print(mod)
}
