% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{resid.linreg}
\alias{resid.linreg}
\title{Residuals method for linreg objects}
\usage{
\method{resid}{linreg}(object, ...)
}
\arguments{
\item{object}{A linreg object}

\item{...}{Additional arguments}
}
\description{
Residuals method for linreg objects
}
