% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Residuals of a linreg object}
\usage{
\method{residuals}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{linreg}.}

\item{...}{Further arguments (not used).}
}
\value{
A numeric vector of residuals.
}
\description{
Returns the vector of residuals \eqn{\hat{e} = y - \hat{y}}.
Called by both \code{resid()} and \code{residuals()}.
}
\examples{
mod <- linreg(Petal.Length ~ Species, data = iris)
head(resid(mod))
}
