% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Coefficients method for linreg objects}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{A linreg object}

\item{...}{Additional arguments}
}
\description{
Coefficients method for linreg objects
}
