% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear regression via QR decomposition (S3)}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{model formula}

\item{data}{data.frame}
}
\value{
an object of class 'linreg'
}
\description{
Linear regression via QR decomposition (S3)
}
