% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression Function}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A formula object (e.g., y ~ x1 + x2).}

\item{data}{A data frame containing the variables.}
}
\value{
An object of class "LinregClass" (RC) with regression results.
}
\description{
Fits a linear regression model using OLS (Ordinary Least Squares).
Returns an RC (Reference Class) object of class "LinregClass".
}
\examples{
data(iris)
mod <- linreg(Petal.Length ~ Species, iris)
}
