% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Diagnostic plots for linreg (ggplot2)}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{a linreg object}

\item{...}{unused; for S3 compatibility}
}
\description{
Diagnostic plots for linreg (ggplot2)
}
