% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot method for linreg objects
Produces Residuals vs Fitted and Scale-Location plots.}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class linreg}

\item{...}{Additional arguments}
}
\description{
Plot method for linreg objects
Produces Residuals vs Fitted and Scale-Location plots.
}
