% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Predicted/fitted values for linreg (method)}
\usage{
\method{pred}{linreg}(object, ...)
}
\arguments{
\item{object}{a linreg object}

\item{...}{unused; for S3 compatibility}
}
\description{
Predicted/fitted values for linreg (method)
}
