% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Predicted/fitted values for linreg objects}
\usage{
\method{pred}{linreg}(object, ...)
}
\arguments{
\item{object}{a model object}

\item{...}{unused; for S3 compatibility}
}
\value{
A numeric vector of fitted values (unnamed).
}
\description{
Predicted/fitted values for linreg objects
}
