% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Prediction method for linreg objects}
\usage{
\method{pred}{linreg}(object, ...)
}
\arguments{
\item{object}{A linreg object}

\item{...}{Additional arguments}
}
\description{
Prediction method for linreg objects
}
