% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pred}
\alias{pred}
\title{Predicted/fitted values for linreg}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{a linreg object}

\item{...}{unused; for S3 compatibility}
}
\description{
Predicted/fitted values for linreg
}
