% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pred}
\alias{pred}
\title{Predicted/fitted values generic}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{a model object}

\item{...}{unused; for S3 compatibility}
}
\value{
Method dispatches to class-specific implementations.
}
\description{
Predicted/fitted values generic
}
