% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.linreg}
\alias{print.linreg}
\title{Print method for linreg}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{a linreg object}

\item{...}{unused; for S3 compatibility}
}
\description{
Print method for linreg
}
