% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{resid.linreg}
\alias{resid.linreg}
\title{Residuals for linreg}
\usage{
\method{resid}{linreg}(object, ...)
}
\arguments{
\item{object}{a linreg object}

\item{...}{unused; for S3 compatibility}
}
\value{
A numeric vector of residuals (unnamed).
}
\description{
Residuals for linreg
}
