% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Residuals for linreg (method for residuals())}
\usage{
\method{residuals}{linreg}(object, ...)
}
\arguments{
\item{object}{a linreg object}

\item{...}{unused; for S3 compatibility}
}
\description{
Residuals for linreg (method for residuals())
}
