% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summary for linreg}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{a linreg object}

\item{...}{unused; for S3 compatibility}
}
\value{
A data.frame with columns: Estimate, Std. Error, t value, Pr(>|t|)
}
\description{
Summary for linreg
}
